Making contract resolver compile with latest Json.NET.
Per https://github.com/elastic/elasticsearch-net/issues/3020, the constructor that was removed from Json.NET appears to have been one that cached contracts -- that is, it allowed the caller to choose whether to use shared contracts or initialize them for every resolver.

The problem, though, is that by dropping the boolean constructor, the shared cache is now lost, and so each resolver will have to introspect the contracts again and again.

One of the notes in #elastic/elasticsearch-net/3020 mentioned that in 5.x the Elastic team modified the resolver to cache contracts per ConnectionSettings, which probably makes the most sense.

I have implemented [this proposed solution](https://github.com/elastic/elasticsearch-net/issues/3020#issuecomment-356475744) in order to mitigate the problem, although I do not know the severity of how this will impact performance.